+++ panel/app/Enums/AdminLogAction.php
@@ -22,6 +22,8 @@ enum AdminLogAction: string
     case TunnelEgressIpUpdated = 'tunnel_egress_ip_updated';
     case VlanProvisioned = 'vlan_provisioned';
     case VlanDeprovisioned = 'vlan_deprovisioned';
+    case BackupCreated = 'backup_created';
+    case BackupDeleted = 'backup_deleted';
 
     public function label(): string
     {
@@ -37,13 +39,15 @@ enum AdminLogAction: string
             self::TunnelEgressIpUpdated => 'Tunnel egress IP updated',
             self::VlanProvisioned => 'VLAN added',
             self::VlanDeprovisioned => 'VLAN deleted',
+            self::BackupCreated => 'Backup created',
+            self::BackupDeleted => 'Backup deleted',
         };
     }
 
     public function color(): string
     {
         return match ($this) {
-            self::AdminLoginFailed, self::PpskDeleted, self::VlanDeprovisioned => 'danger',
+            self::AdminLoginFailed, self::PpskDeleted, self::VlanDeprovisioned, self::BackupDeleted => 'danger',
             self::PpskDisabled => 'warning',
             default => 'success',
         };
@@ -66,6 +70,8 @@ enum AdminLogAction: string
             self::TunnelEgressIpUpdated => 'heroicon-o-globe-alt',
             self::VlanProvisioned => 'heroicon-o-rectangle-group',
             self::VlanDeprovisioned => 'heroicon-o-trash',
+            self::BackupCreated => 'heroicon-o-archive-box',
+            self::BackupDeleted => 'heroicon-o-trash',
         };
     }
 }